+++ system/modules/Monitoring/languages/de/tl_monitoring.php
@@ -40,14 +40,6 @@ $GLOBALS['TL_LANG']['tl_monitoring']['test_string']                 = array('Tes
 $GLOBALS['TL_LANG']['tl_monitoring']['disable']                     = array('Deaktivieren', 'Den Monitoring Eintrag vorübergehend deaktivieren.');
 $GLOBALS['TL_LANG']['tl_monitoring']['last_test_date']              = array('Letzter Test', 'Das Datum des letzten Tests.');
 $GLOBALS['TL_LANG']['tl_monitoring']['last_test_status']            = array('Letzter Status', 'Der Status des letzten Tests.');
-$GLOBALS['TL_LANG']['tl_monitoring']['additional_info_actuality']   = array('Aktualität', 'Geben Sie zu diesem Monitoring Eintrag was aktuell oder veraltet ist.');
-$GLOBALS['TL_LANG']['tl_monitoring']['additional_info_contact']     = array('Kontakt Informationen', 'Geben Sie die Daten zur Kontaktaufname für diesen Monitoring Eintrag an.');
-$GLOBALS['TL_LANG']['tl_monitoring']['additional_info_contao']      = array('Contao Konfiguration', 'Geben Sie die Konfigurationsdaten für Contao zu diesen Monitoring Eintrag an.');
-$GLOBALS['TL_LANG']['tl_monitoring']['additional_info_maintenance'] = array('Wartungskonfiguration', 'Geben Sie die Daten zur Wartungskonfiguration für diesen Monitoring Eintrag an.');
-$GLOBALS['TL_LANG']['tl_monitoring']['additional_info_system']      = array('Systemkonfiguration', 'Geben Sie die Daten zum Web System für diesen Monitoring Eintrag an.');
-$GLOBALS['TL_LANG']['tl_monitoring']['additional_info_field']       = array('Feld', 'Wählen Sie ein Feld aus.');
-$GLOBALS['TL_LANG']['tl_monitoring']['additional_info_actual']      = array('Ja oder nein', 'Geben Sie an ob die Feldbedingung zutrifft.');
-$GLOBALS['TL_LANG']['tl_monitoring']['additional_info_value']       = array('Wert', 'Geben Sie einen Wert für das Feld ein.');
 
 /**
  * Legenden
@@ -55,7 +47,6 @@ $GLOBALS['TL_LANG']['tl_monitoring']['additional_info_value']       = array('Wer
 $GLOBALS['TL_LANG']['tl_monitoring']['website_legend']         = 'Webseite';
 $GLOBALS['TL_LANG']['tl_monitoring']['test_legend']            = 'Testen';
 $GLOBALS['TL_LANG']['tl_monitoring']['last_test_legend']       = 'Letztes Testergebnis';
-$GLOBALS['TL_LANG']['tl_monitoring']['additional_info_legend'] = 'Zusätzliche Informationen';
 
 /**
  * References